refacto(#7): group fields into classes
safetynetalert/safetynetalertsapi/model :

 - Create class Identity to group common fields: firstName and lastName;
 - Create class Address to group fields: address, city and zip;
 - Update fields type in MedicalRecord and Person to reflect the new
structure.
 - Add missing @Override annotations on toString() methods.

Relates to #7
