+++ src/main/java/com/safetynet/safetynetalertsapi/model/Identity.java
@@ -1,0 +1,30 @@
+package com.safetynet.safetynetalertsapi.model;
+
+public class Identity {
+	private String firstName;
+	
+	private String lastName;
+
+	public String getFirstName() {
+		return firstName;
+	}
+
+	public void setFirstName(String firstName) {
+		this.firstName = firstName;
+	}
+
+	public String getLastName() {
+		return lastName;
+	}
+
+	public void setLastName(String lastName) {
+		this.lastName = lastName;
+	}
+	
+	@Override
+	public String toString() {
+		return "firstname : " + this.firstName
+				+ " \n lastname : " + this.lastName;
+		
+	}
+}
